Added cairo1 rpc hint processor to test
