+++ cairo-oracle-hint-processor/Cargo.toml
@@ -6,14 +6,27 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+anyhow = { workspace = true }
+bincode = { version = "2.0.0-rc.3", default-features = false, features = [
+    "serde",
+] }
 cairo-lang-casm = { workspace = true }
+cairo-lang-sierra = { workspace = true }
+cairo-lang-sierra-to-casm = { workspace = true }
+cairo-lang-sierra-type-size = { workspace = true }
 cairo-lang-runner = { workspace = true }
 cairo-lang-utils = { workspace = true }
 cairo-proto-serde = { workspace = true }
 cairo-vm = { workspace = true }
+camino = { workspace = true }
+clap = { workspace = true }
+itertools = { workspace = true }
 num-bigint = { workspace = true }
 num-integer = { workspace = true }
 num-traits = { workspace = true }
 reqwest = { workspace = true }
+scarb-metadata = { workspace = true }
+scarb-ui = { workspace = true }
 serde_json = { workspace = true }
 starknet-types-core = { workspace = true }
+thiserror = { workspace = true }
